+++ cypress/support/cypress-util.ts
@@ -1,5 +1,4 @@
 import {Util} from '../../src/app/lib/util';
-import {ExtensionDefs} from "../../src/app/lib/extension-defs";
 import {JsonPointer} from "json-ptr";
 import {format, parseISO} from 'date-fns';
 
@@ -212,6 +211,28 @@ export class CypressUtil {
     });
   };
 
+  /**
+   * Recursively removes a specified field from an object or an array of objects.
+   *
+   * @param obj - The input object or array of objects from which the specified field should
+   *              be removed.
+   * @param field - A string representing the name of the field to be omitted from the object(s).
+   * @returns - A new object or array of objects with the specified field removed.
+   */
+  static omitField(obj, field) {
+    if (Array.isArray(obj)) {
+      return obj.map(item => this.omitField(item, field));
+    } else if (obj && typeof obj === 'object') {
+
+      const { [field]: _, ...newObj } = obj;
+      for (const key in newObj) {
+        newObj[key] = this.omitField(newObj[key], field);
+      }
+      return newObj;
+    }
+    return obj;
+  }
+ 
   /**
    * Capture clipboard content.
    *