+++ src/app/app.module.ts
@@ -40,7 +40,6 @@ import { AppFormElementComponent } from './lib/widgets/form-element/form-element
 import { TitleComponent } from './lib/widgets/title/title.component';
 import { ElementChooserComponent } from './lib/widgets/element-chooser/element-chooser.component';
 import { AutoCompleteComponent } from './lib/widgets/auto-complete/auto-complete.component';
-import { RadioComponent } from './lib/widgets/radio/radio.component';
 import { HeaderComponent } from './header/header.component';
 import { FooterComponent } from './footer/footer.component';
 import {MatSelectModule} from '@angular/material/select';
@@ -89,6 +88,7 @@ import { DateComponent } from './lib/widgets/date/date.component';
 import { TextAreaComponent } from './lib/widgets/textarea/textarea.component';
 import {DatetimeComponent} from './lib/widgets/datetime/datetime.component';
 import {LabelComponent} from './lib/widgets/label/label.component';
+import { EditableLinkIdComponent } from './lib/widgets/editable-link-id/editable-link-id.component';
 import {CdkCopyToClipboard} from "@angular/cdk/clipboard";
 import {CodemirrorModule} from "@ctrl/ngx-codemirror";
 import {HelpTextComponent} from "./lib/widgets/help-text/help-text.component";
@@ -109,7 +109,6 @@ import {HelpTextComponent} from "./lib/widgets/help-text/help-text.component";
     AppFormElementComponent,
     TitleComponent,
     ElementChooserComponent,
-    RadioComponent,
     HeaderComponent,
     FooterComponent,
     EnableWhenSourceComponent,
@@ -156,6 +155,7 @@ import {HelpTextComponent} from "./lib/widgets/help-text/help-text.component";
     DateComponent,
     TextAreaComponent,
     DatetimeComponent,
+    EditableLinkIdComponent,
     HelpTextComponent
   ],
   imports: [