+++ src/app/lib/widgets/restrictions-operator/restrictions-operator.component.ts
@@ -1,10 +1,8 @@
 import {Component, ElementRef, EventEmitter, OnInit, Output, ViewChild} from '@angular/core';
 import {SelectComponent} from '../select/select.component';
 import {RestrictionsComponent} from '../restrictions/restrictions.component';
-import fhir from 'fhir/r4';
 import {RestrictionOperatorService} from '../../../services/restriction-operator.service';
 import {fhirPrimitives} from '../../../fhir';
-import { FormService } from 'src/app/services/form.service';
 
 /**
  * Used to get acceptance from parent component. The parent should subscribe to service