+++ src/app/services/form.service.ts
@@ -3,7 +3,7 @@
  */
 import {inject, Injectable, SimpleChange} from '@angular/core';
 import {IDType, ITreeNode} from '@bugsplat/angular-tree-component/lib/defs/api';
-import {TreeModel} from '@bugsplat/angular-tree-component';
+import {TreeModel, TreeNode} from '@bugsplat/angular-tree-component';
 import fhir from 'fhir/r4';
 import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
 import {MessageDlgComponent, MessageType} from '../lib/widgets/message-dlg/message-dlg.component';
@@ -30,14 +30,39 @@ import {GuidingStep, Util} from '../lib/util';
 import {FetchService} from './fetch.service';
 import {TerminologyServerComponent} from '../lib/widgets/terminology-server/terminology-server.component';
 import {ExtensionsService} from './extensions.service';
+import { LiveAnnouncer } from '@angular/cdk/a11y';
 
 declare var LForms: any;
+
+export interface ErrorNode {
+  [fieldName:string]: any [];
+}
+
+export interface TreeNodeStatus{
+  treeNodeId: string;
+  linkId: string;
+  hasError?: boolean;
+  childHasError?: boolean;
+  errorMessage?: string;
+  errors?: ErrorNode;
+}
+
+export type TreeNodeStatusMap = {
+  [key:string]: TreeNodeStatus;
+}
+
+export type LinkIdTrackerMap = {
+  [linkIdKey:string] : string[];
+}
 
 @Injectable({
   providedIn: 'root'
 })
 export class FormService {
   static _lformsLoaded$ = new Subject<string>();
+  static readonly TREE_NODE_ID = "__$treeNodeId";
+  _validationStatusChanged$: Subject<void> = new Subject<void>();
+
 
   private _loading = false;
   _guidingStep$: Subject<GuidingStep> = new Subject<GuidingStep>();
@@ -62,7 +87,50 @@ export class FormService {
   fetchService = inject(FetchService);
   formLevelExtensionService = inject(ExtensionsService);
 
-  constructor(private modalService: NgbModal, private http: HttpClient) {
+  treeNodeStatusMap: TreeNodeStatusMap;
+  linkIdTracker:LinkIdTrackerMap = {};
+
+  // All operators
+  operatorOptions: any [] = [
+    {option: 'exists', label: 'Not empty'},
+    {option: 'notexists', label: 'Empty'},
+    {option: '=', label: '='},
+    {option: '!=', label: '!='},
+    {option: '>', label: '>'},
+    {option: '<', label: '<'},
+    {option: '>=', label: '>='},
+    {option: '<=', label: '<='}
+  ];
+
+  // A subset of operators for certain types
+  operatorOptions2: any [] = this.operatorOptions.filter((e) => {
+    return (
+      e.option === 'exists' ||
+      e.option === 'notexists' ||
+      e.option === '=' ||
+      e.option === '!='
+    );
+  });
+
+  // Operators based on type.
+  enableWhenOperatorOptions = {
+    decimal: this.operatorOptions,
+    integer: this.operatorOptions,
+    quantity: this.operatorOptions,
+    date: this.operatorOptions,
+    dateTime: this.operatorOptions,
+    time: this.operatorOptions,
+    string: this.operatorOptions,
+    text: this.operatorOptions,
+    url: this.operatorOptions2,
+    boolean: this.operatorOptions2,
+    choice: this.operatorOptions2,
+    'open-choice': this.operatorOptions2,
+    attachment: this.operatorOptions2,
+    reference: this.operatorOptions2
+  };
+
+  constructor(private modalService: NgbModal, private http: HttpClient, private liveAnnouncer: LiveAnnouncer ) {
     [{schema: ngxItemSchema as any, layout: itemLayout}, {schema: ngxFlSchema as any, layout: flLayout}].forEach((obj) => {
       if(!obj.schema.definitions) {
         obj.schema.definitions = {};
@@ -190,6 +258,13 @@ export class FormService {
   }
 
   /**
+   * Getter for validation status changed Observable
+   */
+  get validationStatusChanged$(): Observable<void> {
+    return this._validationStatusChanged$.asObservable();
+  }
+
+  /**
    * Getter for form reset Observable
    */
   get formReset$(): Observable<void> {
@@ -255,6 +330,380 @@ export class FormService {
     this._guidingStep$.next(step);
   }
 
+  /**
+   * Traverses through the tree nodes, stores each node in the 'validationNodes'
+   * array, and returns this array to be used for validation during initialization.
+   * @returns - array of TreeNodes.
+   */
+  loadValidationNodes(): TreeNode[] {
+    const validationNodes: TreeNode[] = [];
+    function recurse(node: TreeNode): void {
+      validationNodes.push(node);
+
+      if (node.hasChildren) {
+        for (const child of node.children) {
+          recurse(child);
+        }
+      }
+    }
+
+    const roots = this.treeModel.roots;
+    if (roots) {
+      for (const root of roots) {
+        recurse(root);
+      }
+    }
+
+    return validationNodes;
+  }
+
+  /**
+   * Walk through the treeModel and populate the TreeNodeStatus for each of the
+   * TreeNodes into the TreeNodeStatusMap.
+   */
+  loadTreeNodeStatusMap(): void {
+    const treeNodeStatusMap: TreeNodeStatusMap = {};
+    function recurse(node: TreeNode): void {
+      treeNodeStatusMap[node.data[FormService.TREE_NODE_ID]] = {
+        treeNodeId: node.data[FormService.TREE_NODE_ID],
+        linkId: (node.data.linkId) ? node.data.linkId : ''
+      }
+
+      if (node.hasChildren) {
+        for (const child of node.children) {
+          recurse(child);
+        }
+      }
+    }
+
+    if (!this.treeNodeStatusMap || Object.keys(this.treeNodeStatusMap).length === 0) {
+      const roots = this.treeModel.roots;
+      if (roots) {
+        for (const root of roots) {
+          recurse(root);
+        }
+      }
+
+      this.treeNodeStatusMap = treeNodeStatusMap;
+    }
+  };
+
+  /**
+   * Add Tree Node Status for error tracking when an item is added to the Questionnaire
+   * @param treeNodeId - tree node id
+   * @param linkId - linkId associated with item of the node.
+   */
+  addTreeNodeStatus(treeNodeId: string, linkId: string): void {
+    if (!(treeNodeId in this.treeNodeStatusMap)) {
+      this.treeNodeStatusMap[treeNodeId] = {
+        treeNodeId: treeNodeId,
+        linkId: (linkId) ? linkId : ''
+      }
+    }
+  };
+
+  /**
+   * Remove the Tree Node Status from error tracking when the item is deleted.
+   * @param treeNodeId - tree node id
+   */
+  deleteTreeNodeStatus(treeNodeId: string): void {
+    delete this.treeNodeStatusMap[treeNodeId];
+  }
+
+  /**
+   * The 'linkId' must be unique within the questionnaire. Check if the edited 'linkId'
+   * already exists in the questionnaire using the 'linkIdTracker'.
+   * @param newLinkId - linkId associated with item of the node.
+   * @param treeNodeId - tree node id.
+   * @returns true if the edited 'linkId' is not unique, otherwie return false.
+   */
+  treeNodeHasDuplicateLinkIdByLinkIdTracker(newLinkId: string, treeNodeId: string): boolean {
+    if (!treeNodeId || !this.treeNodeStatusMap)
+      return false;
+
+    return (newLinkId in this.linkIdTracker && ( this.linkIdTracker[newLinkId].length > 1 || (this.linkIdTracker[newLinkId].length === 1 && this.linkIdTracker[newLinkId].indexOf(treeNodeId.toString()) === -1)) );
+  };
+
+  /**
+   * Check if the tree node for the given tree node id contains an error.
+   * @param treeNodeId - tree node id.
+   * @param includeChildNodes - indicates whether to include child nodes in this check.
+   * @returns true if the tree node contains error, otherwise false.
+   */
+  isTreeNodeHasErrorById(treeNodeId: string, includeChildNodes: boolean = true): boolean {
+    if (this.treeNodeStatusMap && this.treeNodeStatusMap[treeNodeId]) {
+      const nodeHasError = ('hasError' in this.treeNodeStatusMap[treeNodeId]) ? this.treeNodeStatusMap[treeNodeId]['hasError'] : false;
+      if (includeChildNodes) {
+        const childNodeHasError = ('childHasError' in this.treeNodeStatusMap[treeNodeId]) ? this.treeNodeStatusMap[treeNodeId]['childHasError'] : false;
+        return (nodeHasError || childNodeHasError);
+      } else {
+        return nodeHasError;
+      }
+    }
+    return false;
+  }
+
+  /**
+   * Check if the focus node contains an error.
+   * @returns true if the focus node contains error, otherwise false.
+   */
+  isFocusNodeHasError(): boolean {
+    if (this.treeModel) {
+      const node = this.treeModel.getFocusedNode();
+      if (node)
+        return this.isTreeNodeHasErrorById(node.data[FormService.TREE_NODE_ID], false);
+    }
+    return false;
+  }
+
+  /**
+   * Return the TreeNodeStatus for the given tree node id.
+   * @param treeNodeId - tree node id.
+   * @returns - TreeNodeStatus object if found, otherwise null.
+   */
+  getTreeNodeStatusById(treeNodeId: string): TreeNodeStatus | null {
+    if (this.treeNodeStatusMap && (treeNodeId in this.treeNodeStatusMap)) {
+      return this.treeNodeStatusMap[treeNodeId];
+    }
+    return null;
+  }
+
+  /**
+   * Check if the sibling nodes contain errors.
+   * @param node - current tree node
+   * @returns true if any of the sibling nodes contain errors, otherwise false.
+   */
+  siblingHasError(node:ITreeNode): boolean {
+    let siblingHasError = false;
+    if (node.parent && !node.isRoot && node.parent.hasChildren) {
+      siblingHasError = node.parent.children.some((n) => {
+        const siblingIdStr = n.data[FormService.TREE_NODE_ID];
+        return node.data[FormService.TREE_NODE_ID] !== siblingIdStr && (this.treeNodeStatusMap[siblingIdStr]?.hasError ?? false);
+      });
+    }
+    return siblingHasError;
+  }
+
+  /**
+   * Set the ancestor nodes' 'childHasError' status to 'false'.
+   * @param node - Ancestor node.
+   */
+  removeErrorFromAncestorNodes(node: ITreeNode): void {
+    const nodeIdStr = node.data[FormService.TREE_NODE_ID];
+    if (nodeIdStr in this.treeNodeStatusMap) {
+      this.treeNodeStatusMap[nodeIdStr]['childHasError'] = false;
+    }
+
+    if (node.parent && !node.isRoot &&
+        (!('childHasError' in this.treeNodeStatusMap[nodeIdStr]) || !this.treeNodeStatusMap[nodeIdStr]['childHasError'])) {
+
+      const siblingHasError = this.siblingHasError(node);
+      if (!siblingHasError)
+        this.removeErrorFromAncestorNodes(node.parent);
+    }
+  }
+
+  /**
+   * Set the ancestor nodes' 'childHasError' status to 'true'.
+   * @param node - Ancestor node.
+   */
+  addErrorForAncestorNodes(node: ITreeNode): void {
+    const nodeIdStr = node.data[FormService.TREE_NODE_ID];
+
+    if (nodeIdStr in this.treeNodeStatusMap) {
+      this.treeNodeStatusMap[nodeIdStr]['childHasError'] = true;
+    }
+
+    // The root node may have parent, but it is not an item
+    if (node.parent && !node.isRoot) {
+      this.addErrorForAncestorNodes(node.parent);
+    }
+  }
+
+  /**
+   * Update validation status to the TreeNodeStatusMap.
+   * @param treeNodeId - tree node id
+   * @param linkId - linkId associated with item of the node.
+   * @param fieldName - name of the field being validated.
+   * @param errors - Null, if the validation passes. Set the 'hasError' status to 'false'
+   *                 and the ancestor nodes' 'chilcHasError' status to 'false'.
+   *                 Array of errors if the the validation fails. Set the 'hasError' status
+   *                 to 'true' and the ancestor nodes' 'childHasError' status to 'true'.
+   *                 Also stores the array of errors objects.
+   */
+  updateValidationStatus(treeNodeId: string, linkId: string, fieldName: string, errors: any[]): void {
+    if (!this.treeNodeStatusMap || !this.treeNodeStatusMap[treeNodeId])
+      return;
+
+    this.treeNodeStatusMap[treeNodeId]['linkId'] = linkId;
+    if (errors) {
+      if (!this.treeNodeStatusMap[treeNodeId]['errors'])
+        this.treeNodeStatusMap[treeNodeId]['errors'] = {};
+      this.treeNodeStatusMap[treeNodeId]['errors'][fieldName] = errors;
+      this.treeNodeStatusMap[treeNodeId]['hasError'] = true;
+    } else {
+      if (this.treeNodeStatusMap[treeNodeId]['errors'] && fieldName in this.treeNodeStatusMap[treeNodeId]['errors']) {
+        this.liveAnnouncer.announce('Error is resolved for this node.');
+        delete this.treeNodeStatusMap[treeNodeId]['errors'][fieldName];
+      }
+
+      this.treeNodeStatusMap[treeNodeId]['hasError'] = (Object.keys(this.treeNodeStatusMap[treeNodeId]?.errors ?? {}).length > 0);
+    }
+
+    const node = this.getTreeNodeById(treeNodeId);
+
+    if (node && node.parent && !node.isRoot) {
+      if (errors) {
+        this.addErrorForAncestorNodes(node.parent);
+
+      } else {
+        // Only remove the error from ancestor nodes if
+        // - the focus node does not contain errors
+        // - the child nodes do not contain errors
+        // - the sibling nodes do not contain errors
+        const siblingHasError = this.siblingHasError(node);
+        if (!siblingHasError && Object.keys(this.treeNodeStatusMap[treeNodeId]?.errors ?? {}).length === 0 &&
+            (!('childHasError' in this.treeNodeStatusMap[treeNodeId]) || !this.treeNodeStatusMap[treeNodeId]['childHasError']))
+          this.removeErrorFromAncestorNodes(node.parent);
+      }
+    }
+  }
+
+  /**
+   * Delete the enableWhen error that matches the given rowIndex and adjusts the indices of the
+   * remaining enableWhen error keys that are greater than the rowIndex from the TreeNodeStatusMap.
+   *
+   * The errors are tracked using the key format 'enableWhen_' + <row index>.  When an enableWhen
+   * is deleted, the existing enableWhen errors may need to be shifted up accordingly.
+   *
+   * @param treeNodeId - tree node id.
+   * @param rowIndex - the row index of the Enable When condition that is being deleted.
+   */
+  deleteErrorAndAdjustEnableWhenIndexes(treeNodeId: string, rowIndex: number): void {
+    if (!this.treeNodeStatusMap || !this.treeNodeStatusMap[treeNodeId])
+      return;
+
+    const fieldName = `enableWhen_${rowIndex}`;
+    if (this.treeNodeStatusMap[treeNodeId]['errors'] && fieldName in this.treeNodeStatusMap[treeNodeId]['errors']) {
+      delete this.treeNodeStatusMap[treeNodeId]['errors'][fieldName];
+    }
+
+    const enableWhenKeys = Object.keys(this.treeNodeStatusMap[treeNodeId]['errors']);
+
+    enableWhenKeys.forEach(ewKey => {
+      const match = ewKey.match(/enableWhen_(\d+)/);
+      const keyIndex = match ? Number(match[1]) : -1;
+
+      if (!isNaN(keyIndex) && keyIndex > rowIndex) {
+        this.treeNodeStatusMap[treeNodeId]['errors'][`enableWhen_${keyIndex - 1}`] =
+          this.treeNodeStatusMap[treeNodeId]['errors'][`enableWhen_${keyIndex}`];
+        delete this.treeNodeStatusMap[treeNodeId]['errors'][`enableWhen_${keyIndex}`];
+      }
+    });
+
+    this.treeNodeStatusMap[treeNodeId]['hasError'] = (Object.keys(this.treeNodeStatusMap[treeNodeId]?.errors ?? {}).length > 0);
+    this._validationStatusChanged$.next(null);
+  }
+
+
+  /**
+   * Remove treeNodeStatusMap from local storage.
+   */
+  clearAutoSavedTreeNodeStatusMap() {
+    localStorage.removeItem('treeMap');
+    this.treeNodeStatusMap = {};
+  }
+
+  /**
+   * Load and initialize the 'linkIdTracker' to track duplicate
+   * 'linkIds' for each tree node.
+   */
+  loadLinkIdTracker(): void {
+    Object.values(this.treeNodeStatusMap).map((node) => {
+      const linkId = node.linkId;
+      const nodeId = node.treeNodeId;
+
+      if (linkId in this.linkIdTracker) {
+        if (this.linkIdTracker[linkId].indexOf(nodeId) === -1)
+          this.linkIdTracker[linkId].push(nodeId);
+      } else {
+        this.linkIdTracker[linkId] = [nodeId];
+      }
+    });
+  }
+
+  /**
+   * Retrieve tree node ids for a specified 'linkId'.
+   * @param linkId - linkId associated with item of the node.
+   * @returns Array of tree node id(s) if found, otherwise null
+   */
+  getNodeIdsByLinkId(linkId: string): string[] | null {
+    return ((linkId || linkId === '') && linkId in this.linkIdTracker) ? this.linkIdTracker[linkId] : null;
+  }
+
+  /**
+   * Check if the linkId exists in the Questionnaire/linkIdTracker.
+   * @param linkId - linkId associated with item of the node.
+   * @returns True if the linkId exists in the linkIdTracker.
+   */
+  isValidLinkId(linkId: string): boolean {
+    return !!this.getNodeIdsByLinkId(linkId);
+  }
+
+  /**
+   * Add the 'linkId' and associated tree node id to the 'linkIdTracker'.
+   * @param treeNodeId - tree node id.
+   * @param linkId - linkId associated with item of the node.
+   */
+  addLinkIdToLinkIdTracker(treeNodeId: string, linkId: string): void {
+    if (linkId in this.linkIdTracker) {
+      if (this.linkIdTracker[linkId].indexOf(treeNodeId) === -1)
+        this.linkIdTracker[linkId].push(treeNodeId);
+    } else {
+      this.linkIdTracker[linkId] = [treeNodeId];
+    }
+  }
+
+  /**
+   * Remove the specified 'linkId' from the 'linkIdTracker', but if the 'linkId'
+   * corresponds to an array containing multiple ids, only remove the matching
+   * 'treeNodeId' from the array.
+   * @param treeNodeId - tree node id.
+   * @param linkId - linkId associated with item of the node.
+   */
+  removeLinkIdFromLinkIdTracker(treeNodeId: string, linkId: string): void {
+    if (linkId in this.linkIdTracker) {
+      if (this.linkIdTracker[linkId].length > 1) {
+        const index = this.linkIdTracker[linkId].indexOf(treeNodeId);
+        if (index > -1) {
+          this.linkIdTracker[linkId].splice(index, 1);
+        }
+      } else {
+        delete this.linkIdTracker[linkId];
+      }
+    }
+  }
+
+  /**
+   * Update the 'linkIdTracker' whenever the 'linkId' value changes.
+   * @param prevLinkId - existing linkId associated with item of the node.
+   * @param newLinkId - updated linkId associated with item of the node.
+   * @param treeNodeId - tree node id
+   */
+  updateLinkIdForLinkIdTracker(prevLinkId: string, newLinkId: string, treeNodeId: string): void {
+    if (treeNodeId) {
+      if (prevLinkId)
+        this.removeLinkIdFromLinkIdTracker(treeNodeId, prevLinkId);
+      this.addLinkIdToLinkIdTracker(treeNodeId, newLinkId);
+    }
+  }
+
+  /**
+   * Reset the linkIdTracker
+   */
+  clearLinkIdTracker(): void {
+    this.linkIdTracker = {};
+  }
 
   /**
    * Intended to collect source items for enable when logic
@@ -269,6 +718,14 @@ export class FormService {
     return ret;
   }
 
+  /**
+   * Returns a list of EnableWhen operators based on the given answer type.
+   * @param answerType - Type of the source item.
+   * @returns - a list of operators associated with the given answer type.
+   */
+  getEnableWhenOperatorListByAnswerType(answerType: string): any [] {
+    return this.enableWhenOperatorOptions[answerType];
+  }
 
   /**
    * Get sources excluding the branch of a given node.
@@ -319,11 +776,11 @@ export class FormService {
 
 
   /**
-   * Get node by its id.
-   * @param id
+   * Get node by its tree node id.
+   * @param treeNodeId
    */
-  getTreeNodeById(id: IDType): ITreeNode {
-    return this.treeModel.getNodeById(id);
+  getTreeNodeById(treeNodeId: IDType): ITreeNode {
+    return this.treeModel.getNodeById(treeNodeId);
   }
 
 
@@ -332,9 +789,8 @@ export class FormService {
    * @param linkId
    */
   getTreeNodeByLinkId(linkId: string): ITreeNode {
-    return this.findNodeByLinkId(this.treeModel.roots, linkId);
+    return this.findNodeByLinkId(this.treeModel?.roots, linkId);
   }
-
 
   /**
    * Checks if the focused node has an extension.
@@ -350,7 +806,7 @@ export class FormService {
    * @returns True if the focused node's data contains sub-items. Otherwise false.
    */
   hasSubItems(): boolean {
-    return !!this.treeModel?.getFocusedNode()?.data?.item ?? false;
+    return !!this.treeModel?.getFocusedNode()?.data?.item;
   }
 
   /**