+++ CHANGELOG.md
@@ -2,9 +2,20 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [9.6.10] 2024-11-20
+## [9.6.12] 2024-11-20
 ### Added
 - Added a message for FHIR server validation.
+
+## [9.6.11] 2024-10-24
+### Fixed
+- Fix bug in units when adding a number type question from LOINC.
+### Changed
+- Added aria-labels to radio group elements to improve accessibility.
+
+## [9.6.10] 2024-11-01
+### Added
+- Added functionality for item validation and corresponding error display.
+- Added editable link id field.
 
 ## [9.6.9] 2024-10-17
 ### Changed